Use user-error instead of error where applicable
Quoting from section 11.7.3.1 How to Signal an Error in the GNU Emacs
Lisp Reference Manual (version 26.2):

"[user-error] behaves exactly like error, except that it uses the error
symbol user-error rather than error. [It] is intended to report errors
on the part of the user, rather than errors in the code itself. For
example, if you try to use the command Info-history-back (l) to move
back beyond the start of your Info browsing history, Emacs signals a
user-error. Such errors do not cause entry to the debugger, even when
debug-on-error is non-nil."